requirements.txt
# streamlit==1.52.0

# in app.py
numpy
flask
Flask-Cors
google-generativeai
supabase
requests
xgboost-cpu
statsmodels
statsforecast
joblib
google-cloud-storage
